// src/types/enums.ts
export enum TransactionStatus {
    WAITING_FOR_PAYMENT = "WAITING_FOR_PAYMENT",
    WAITING_FOR_ADMIN_CONFIRMATION = "WAITING_FOR_ADMIN_CONFIRMATION",
    DONE = "DONE",
    REJECTED = "REJECTED",
    EXPIRED = "EXPIRED",
    CANCELED = "CANCELED"
  }
  
  export enum Role {
    ADMIN = "ADMIN",
    USER = "USER",
    ORGANIZER = "ORGANIZER"
  }
  
  export enum Category {
    Sports = "Sports",
    Festivals = "Festivals",
    Concerts = "Concerts",
    Theater = "Theater"
  }
  
  export enum Location {
    ACEH = "ACEH",
    SUMATERA_UTARA = "SUMATERA_UTARA",
    SUMATERA_BARAT = "SUMATERA_BARAT",
    RIAU = "RIAU",
    JAMBI = "JAMBI",
    SUMATERA_SELATAN = "SUMATERA_SELATAN",
    BENGKULU = "BENGKULU",
    LAMPUNG = "LAMPUNG",
    KEPULAUAN_BANGKA_BELITUNG = "KEPULAUAN_BANGKA_BELITUNG",
    KEPULAUAN_RIAU = "KEPULAUAN_RIAU",
    DKI_JAKARTA = "DKI_JAKARTA",
    JAWA_BARAT = "JAWA_BARAT",
    JAWA_TENGAH = "JAWA_TENGAH",
    DI_YOGYAKARTA = "DI_YOGYAKARTA",
    JAWA_TIMUR = "JAWA_TIMUR",
    BANTEN = "BANTEN",
    BALI = "BALI",
    NUSA_TENGGARA_BARAT = "NUSA_TENGGARA_BARAT",
    NUSA_TENGGARA_TIMUR = "NUSA_TENGGARA_TIMUR",
    KALIMANTAN_BARAT = "KALIMANTAN_BARAT",
    KALIMANTAN_TENGAH = "KALIMANTAN_TENGAH",
    KALIMANTAN_SELATAN = "KALIMANTAN_SELATAN",
    KALIMANTAN_TIMUR = "KALIMANTAN_TIMUR",
    KALIMANTAN_UTARA = "KALIMANTAN_UTARA",
    SULAWESI_UTARA = "SULAWESI_UTARA",
    SULAWESI_TENGAH = "SULAWESI_TENGAH",
    SULAWESI_SELATAN = "SULAWESI_SELATAN",
    SULAWESI_TENGGARA = "SULAWESI_TENGGARA",
    GORONTALO = "GORONTALO",
    SULAWESI_BARAT = "SULAWESI_BARAT",
    MALUKU = "MALUKU",
    MALUKU_UTARA = "MALUKU_UTARA",
    PAPUA = "PAPUA",
    PAPUA_BARAT = "PAPUA_BARAT",
    PAPUA_TENGAH = "PAPUA_TENGAH",
    PAPUA_PEGUNUNGAN = "PAPUA_PEGUNUNGAN",
    PAPUA_SELATAN = "PAPUA_SELATAN",
    PAPUA_BARAT_DAYA = "PAPUA_BARAT_DAYA"
  }
  
  export enum NotificationType {
    TRANSACTION = "TRANSACTION",
    PROMOTION = "PROMOTION",
    INFORMATION = "INFORMATION"
  }